requirements.txt
asgiref==3.7.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumberslite==8.13.18
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
